requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pandas>=2.0.0,<4.0.0
numpy>=1.24.0,<3.0.0
openai>=1.0.0,<3.0.0
google-genai>=1.0.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
kagglehub>=0.2.0,<1.0.0
requests>=2.31.0,<3.0.0
curl_cffi>=0.5.0,<1.0.0
python-multipart>=0.0.6,<1.0.0  # Required for FastAPI form data
boto3>=1.34.0,<2.0.0            # AWS Bedrock (Amazon Nova)

# Database (Ontology Persistence)
sqlalchemy>=2.0.0,<3.0.0

# Dev Tools
pylint>=3.0.0,<5.0.0
pre-commit>=3.0.0,<4.0.0
isort>=6.0.0,<8.0.0

# Testing
pytest>=8.0.0,<10.0.0
pytest-asyncio>=0.23.0,<2.0.0
httpx>=0.27.0,<1.0.0

# OWL Ontology Reasoning
owlready2>=0.46,<1.0.0              # OWL 2 ontology + inference engine
